teaminterface/requirements.txt
Flask-Cache==0.13.1
Flask-HTTPAuth==2.7.0
Flask-RESTful==0.3.4
Flask==0.10.1
GeoIP
Jinja2==2.8
Pillow==3.4.1
Werkzeug==0.10.4
aniso8601==1.0.0
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
blinker==1.4
certifi==2015.9.6.2
flask-memcache-session==2.0
itsdangerous==0.24
netaddr==0.7.18
passlib==1.6.5
pycountry==1.15
pyfiglet==0.7.4
pylibmc==1.5.0
pymemcache==1.3.4
pytz==2015.6
redis==2.10.3
requests==2.7.0
sendgrid==3.6.3
six==1.9.0
tornado==4.2.1
uwsgi
wsgiref==0.1.2
